refactor(chat): extract session.go + repl.go, wire chatcommands registry
- Extract chatSession and buildChatSession into session.go (~350 lines)
- Extract REPL loop into repl.go (~165 lines), all commands dispatch via Registry
- Rewrite commands.go to register /exit, /quit, /forget, /memory, /help,
  /remember, /model, /init, /update into chatcommands.Registry
- Add Result type to chatcommands package for control-flow signals (Reply, Quit)
- Simplify chat.go to ~55 lines of top-level wiring only

Fixes upstream review items 3 & 4:
- chatcommands package is now actively used by the TUI REPL
- chat.go is split into focused files with clear responsibilities
